Rename playlistsList to playlistList
