+++ app/src/main/java/com/sebastianvm/musicplayer/ui/library/playlistlist/PlaylistListNavDestination.kt
@@ -1,4 +1,4 @@
-package com.sebastianvm.musicplayer.ui.library.playlists
+package com.sebastianvm.musicplayer.ui.library.playlistlist
 
 import androidx.navigation.NavGraphBuilder
 import com.sebastianvm.musicplayer.ui.navigation.DestinationType
@@ -6,12 +6,12 @@ import com.sebastianvm.musicplayer.ui.navigation.NavigationDelegate
 import com.sebastianvm.musicplayer.ui.navigation.NavigationRoute
 import com.sebastianvm.musicplayer.ui.navigation.screenDestination
 
-fun NavGraphBuilder.playlistsListNavDestination(navigationDelegate: NavigationDelegate) {
+fun NavGraphBuilder.playlistListNavDestination(navigationDelegate: NavigationDelegate) {
     screenDestination<PlaylistListViewModel>(
         destination = NavigationRoute.PlaylistsRoot,
         destinationType = DestinationType.Screen
     ) { viewModel ->
-        PlaylistsListScreen(
+        PlaylistListScreen(
             screenViewModel = viewModel,
             navigationDelegate = navigationDelegate
         )
